+++ app/Resources/js/app.js
@@ -4,15 +4,15 @@
  *
  */
 require('bootstrap');
-require('../../scss/custom.scss');
+require('../scss/custom.scss');
 
-module.exports = function() {
+module.exports = function () {
   /*
    *	Class Bundle App client side
    */
   const App = class App {
     constructor() {
-      let selectorLang = global.document.getElementById("lang");
+      let selectorLang = global.document.getElementById("language");
       if (selectorLang) {
         selectorLang.addEventListener("change", (e) => {
           //window.location.href = "?lang=" + this.value;
@@ -24,13 +24,14 @@ module.exports = function() {
 
     changeLang(query) {
       if (query) {
-        return window.location.href = "?lang=" + query;
+        return window.location.href = "?language=" + query;
       }
       let form = global.document.getElementById("formLang");
       if (form) {
         form.submit();
       }
     }
+
   };
   return new App();
 }();