+++ app/config/routing.js
@@ -6,24 +6,5 @@ module.exports = {
     defaults: {
       controller: "app:app:index"
     }
-  },
-  login: {
-    pattern: "/login/{type}",
-    defaults: {
-      controller: "app:login:login",
-      type: "nodefony"
-    }
-  },
-  logout: {
-    pattern: "/logout",
-    defaults: {
-      controller: "app:login:logout"
-    }
-  },
-  lang: {
-    pattern: "/lang",
-    defaults: {
-      controller: "app:app:lang"
-    }
   }
 };