+++ app/config/webpack.config.js
@@ -3,7 +3,8 @@ const path = require("path");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const webpackMerge = require('webpack-merge');
 
-const context = path.resolve(__dirname, "..", "Resources", "public");
+// Default context <bundle base directory>
+//const context = path.resolve(__dirname, "..", "Resources", "public");
 const public = path.resolve(__dirname, "..", "Resources", "public", "assets");
 const bundleName = path.basename(path.resolve(__dirname, ".."));
 const publicPath = bundleName + "/assets/";
@@ -21,7 +22,8 @@ module.exports = webpackMerge(config, {
   //context: context,
   target: "web",
   entry: {
-    app: ["./Resources/public/js/app.js"]
+    app: ["./Resources/js/app.js"],
+    users: ["./Resources/js/users.js"]
   },
   output: {
     path: public,
@@ -53,7 +55,7 @@ module.exports = webpackMerge(config, {
         loader: "expose-loader?$!expose-loader?jQuery"
       }, {
         test: /jquery\..*\.js/,
-        loader: "imports?$=jquery,jQuery=jquery,this=>window"
+        loader: "imports-loader?$=jquery,jQuery=jquery,this=>window"
       }, {
         test: /\.(sa|sc|c)ss$/,
         use: [
@@ -70,7 +72,7 @@ module.exports = webpackMerge(config, {
           }, {
             loader: 'postcss-loader', // Run post css actions
             options: {
-              plugins: function() { // post css plugins, can be exported to postcss.config.js
+              plugins: function () { // post css plugins, can be exported to postcss.config.js
                 return [
                   require('precss'),
                   require('autoprefixer')